removed the stock next.js page and replaced it with our home page, landing route change d to login, added anchor tags to signup button and the weverifai logo to route to their respective pages , minor change to login button color, added an MVP /post page, we link queries here later, added +UVerifA
