+++ src/app/home.module.css
@@ -1,0 +1,313 @@
+.container {
+  display: flex;
+  flex-direction: column;
+  align-items: center;
+  text-align: center;
+  min-height: 100vh;
+  background: #f9f9f9;
+  color: #333;
+  font-family: "Inter", sans-serif;
+  transition: background 0.3s ease, color 0.3s ease;
+}
+
+.accent {
+  color: #f36100 !important;
+}
+
+.accent2 {
+  color: #007bff !important;
+}
+
+/* DARK MODE */
+.dark {
+  background: #181818;
+  color: #eee;
+}
+
+.header {
+  width: 100%;
+  display: flex;
+  justify-content: space-between;
+  align-items: center;
+  background: white;
+  padding: 15px 40px;
+  box-shadow: 0 4px 10px rgba(0, 0, 0, 0.281);
+  position: fixed;
+  top: 0;
+  z-index: 100;
+  transition: background 0.3s ease;
+}
+
+.dark .header {
+  background: #222;
+  box-shadow: 0 4px 10px rgba(255, 255, 255, 0.05);
+}
+
+.logo {
+  position: absolute;
+  left: 50%;
+  transform: translateX(-50%);
+  flex-grow: 1;
+  text-align: center;
+  font-size: 1.8rem;
+  font-weight: 600;
+  color: #000;
+}
+
+.dark .logo {
+  color: #eee;
+}
+
+.toggleWrapper {
+  display: flex;
+  align-items: center;
+  gap: 5px;
+}
+
+.switch {
+  position: relative;
+  display: inline-block;
+  width: 40px;
+  height: 20px;
+}
+
+.switch input {
+  opacity: 0;
+  width: 0;
+  height: 0;
+}
+
+.slider {
+  position: absolute;
+  cursor: pointer;
+  top: 0;
+  left: 0;
+  right: 0;
+  bottom: 0;
+  background-color: #007bff;
+  transition: 0.4s;
+  border-radius: 20px;
+}
+
+.slider:before {
+  position: absolute;
+  content: "";
+  height: 16px;
+  width: 16px;
+  left: 2px;
+  bottom: 2px;
+  background-color: #ffffff;
+  transition: 0.4s;
+  border-radius: 50%;
+}
+
+input:checked + .slider {
+  background-color: #f36100;
+}
+
+input:checked + .slider:before {
+  transform: translateX(20px);
+}
+
+.navright {
+  display: flex;
+  gap: 20px;
+  padding: 5px 5px;
+  color: #555;
+  text-decoration: none;
+  font-size: 1rem;
+  transition: 0.3s;
+}
+
+.container.dark .navright {
+  color: #bbb;
+}
+
+.navItem {
+  padding: 8px 16px;
+  gap: 10px;
+  color: #555;
+  text-decoration: none;
+  font-size: 1rem;
+  transition: 0.3s;
+}
+
+.container.dark .navItem {
+  color: #bbb;
+}
+
+.navItem:hover {
+  color: #007bff;
+}
+
+.container.dark .navItem:hover {
+  color: #3ea6ff;
+}
+
+.sidebar {
+  position: fixed;
+  left: -250px;
+  top: 0;
+  width: 250px;
+  height: 100%;
+  background: white;
+  box-shadow: 2px 0 10px rgba(0, 0, 0, 0.1);
+  padding-top: 60px;
+  transition: left 0.3s ease-in-out;
+}
+
+.dark .sidebar {
+  background: #222;
+}
+
+.sidebar.open {
+  left: 0;
+}
+
+.closeButton {
+  background: none;
+  border: none;
+  font-size: 1.5rem;
+  cursor: pointer;
+  color: #555;
+  position: absolute;
+  top: 15px;
+  right: 15px;
+}
+
+.dark .closeButton {
+  color: #bbb;
+}
+
+.forum {
+  margin-top: 100px;
+  text-align: center;
+  width: 90%;
+  max-width: 800px;
+}
+
+.forumTitle {
+  font-size: 2rem;
+  font-weight: 700;
+}
+
+.posts {
+  display: flex;
+  flex-direction: column;
+  gap: 20px;
+  margin-top: 20px;
+}
+
+.post {
+  background: white;
+  padding: 15px;
+  border-radius: 8px;
+  box-shadow: 0 4px 10px rgba(0, 0, 0, 0.281);
+  transition: transform 0.2s ease;
+}
+
+.dark .post {
+  background: #222;
+  box-shadow: 0 4px 10px rgba(255, 255, 255, 0.05);
+}
+
+.post:hover {
+  transform: translateY(-3px);
+}
+
+.postHeader {
+  display: flex;
+  align-items: center;
+  gap: 10px;
+}
+
+.userIcon {
+  font-size: 2rem;
+  color: #007bff;
+}
+
+.postInfo {
+  text-align: left;
+}
+
+.postTitle {
+  font-size: 1.2rem;
+  font-weight: 600;
+}
+
+.postMeta {
+  font-size: 0.9rem;
+  color: #666;
+}
+
+.dark .postMeta {
+  color: #bbb;
+}
+
+.postContent {
+  margin-top: 10px;
+  font-size: 1rem;
+  color: #444;
+}
+
+.dark .postContent {
+  color: #ddd;
+}
+
+.postFooter {
+  display: flex;
+  align-items: center;
+  gap: 5px;
+  margin-top: 10px;
+  color: #555;
+}
+
+.dark .postFooter {
+  color: #bbb;
+}
+
+.commentIcon {
+  font-size: 1rem;
+}
+
+.commentCount {
+  font-size: 0.9rem;
+}
+
+.footer {
+  padding: 20px;
+  background: white;
+  color: #666;
+  text-align: center;
+}
+
+.dark .footer {
+  background: #222;
+  color: #bbb;
+}
+
+.createPostBtn {
+  display: flex;
+  align-items: center;
+  gap: 5px;
+  padding: 8px 16px;
+  background: #007bff;
+  color: #fff;
+  border: none;
+  border-radius: 8px;
+  font-size: 1rem;
+  cursor: pointer;
+  transition: background 0.3s ease;
+}
+
+.container.dark .createPostBtn {
+  background: #f36100;
+}
+
+.createPostBtn:hover {
+  background: #f36100;
+}
+
+.container.dark .createPostBtn:hover {
+  background: #3ea6ff;
+}
