+++ src/app/layout.tsx
@@ -13,8 +13,8 @@ const geistMono = Geist_Mono({
 });
 
 export const metadata: Metadata = {
-  title: "Create Next App",
-  description: "Generated by create next app",
+  title: "WeVerifAI",
+  description: "Trust, but VerifAI",
 };
 
 export default function RootLayout({
