+++ src/app/page.tsx
@@ -1,101 +1,125 @@
-import Image from "next/image";
+"use client";
+import { useState, useEffect, JSX } from "react";
+import { useRouter } from "next/navigation";
+import styles from "./home.module.css";
+import { FaBars, FaTimes, FaUserCircle, FaCommentAlt, FaPlus } from "react-icons/fa";
 
-export default function Home() {
+interface Post {
+  id: number;
+  title: string;
+  author: string;
+  content: string;
+  timePosted: string;
+  commentCount: number;
+}
+
+export default function Home(): JSX.Element {
+  const [darkMode, setDarkMode] = useState<boolean>(false);
+  const [menuOpen, setMenuOpen] = useState<boolean>(false);
+  const router = useRouter();
+
+  useEffect(() => {
+    document.body.classList.toggle(styles.dark, darkMode);
+  }, [darkMode]);
+
+  const generateMockPosts = (): Post[] => {
+    return Array.from({ length: 12 }, (_, i) => ({
+      id: i,
+      title: `Sample Forum Post ${i + 1}`,
+      author: `User${i + 1}`,
+      content: "This is a placeholder for a forum post. Here you can write a discussion topic or question.",
+      timePosted: "2 hours ago",
+      commentCount: Math.floor(Math.random() * 30)
+    }));
+  };
+
+  const [posts, setPosts] = useState<Post[]>([]);
+
+  useEffect(() => {
+    setPosts(generateMockPosts());
+  }, []);
+
+  const handleSignUpClick = () => {
+    router.push("/login");
+  };
+
+  const handleCreatePostClick = () => {
+    router.push("/post");
+  };
+
   return (
-    <div className="grid grid-rows-[20px_1fr_20px] items-center justify-items-center min-h-screen p-8 pb-20 gap-16 sm:p-20 font-[family-name:var(--font-geist-sans)]">
-      <main className="flex flex-col gap-8 row-start-2 items-center sm:items-start">
-        <Image
-          className="dark:invert"
-          src="/next.svg"
-          alt="Next.js logo"
-          width={180}
-          height={38}
-          priority
-        />
-        <ol className="list-inside list-decimal text-sm text-center sm:text-left font-[family-name:var(--font-geist-mono)]">
-          <li className="mb-2">
-            Get started by editing{" "}
-            <code className="bg-black/[.05] dark:bg-white/[.06] px-1 py-0.5 rounded font-semibold">
-              src/app/page.tsx
-            </code>
-            .
-          </li>
-          <li>Save and see your changes instantly.</li>
-        </ol>
+    <div className={`${styles.container} ${darkMode ? styles.dark : ""}`}>
+      <header className={styles.header}>
+        <button 
+          className={styles.menuButton} 
+          onClick={() => setMenuOpen(!menuOpen)}
+        >
+          {menuOpen ? <FaTimes /> : <FaBars />}
+        </button>
 
-        <div className="flex gap-4 items-center flex-col sm:flex-row">
-          <a
-            className="rounded-full border border-solid border-transparent transition-colors flex items-center justify-center bg-foreground text-background gap-2 hover:bg-[#383838] dark:hover:bg-[#ccc] text-sm sm:text-base h-10 sm:h-12 px-4 sm:px-5"
-            href="https://vercel.com/new?utm_source=create-next-app&utm_medium=appdir-template-tw&utm_campaign=create-next-app"
-            target="_blank"
-            rel="noopener noreferrer"
-          >
-            <Image
-              className="dark:invert"
-              src="/vercel.svg"
-              alt="Vercel logomark"
-              width={20}
-              height={20}
-            />
-            Deploy now
-          </a>
-          <a
-            className="rounded-full border border-solid border-black/[.08] dark:border-white/[.145] transition-colors flex items-center justify-center hover:bg-[#f2f2f2] dark:hover:bg-[#1a1a1a] hover:border-transparent text-sm sm:text-base h-10 sm:h-12 px-4 sm:px-5 sm:min-w-44"
-            href="https://nextjs.org/docs?utm_source=create-next-app&utm_medium=appdir-template-tw&utm_campaign=create-next-app"
-            target="_blank"
-            rel="noopener noreferrer"
-          >
-            Read our docs
-          </a>
+        <div className={`${styles.sidebar} ${menuOpen ? styles.open : ""}`}>
+          <nav className={styles.navbar}>
+            <button 
+              className={styles.closeButton} 
+              onClick={() => setMenuOpen(false)}
+            >
+              <FaTimes />
+            </button>
+            <a href="#" className={styles.navItem}>General Discussion</a><br />
+            <a href="#" className={styles.navItem}>Announcements</a><br />
+            <a href="#" className={styles.navItem}>Support</a><br />
+            <a href="#" className={styles.navItem}>Off-Topic</a><br />
+          </nav>
         </div>
-      </main>
-      <footer className="row-start-3 flex gap-6 flex-wrap items-center justify-center">
-        <a
-          className="flex items-center gap-2 hover:underline hover:underline-offset-4"
-          href="https://nextjs.org/learn?utm_source=create-next-app&utm_medium=appdir-template-tw&utm_campaign=create-next-app"
-          target="_blank"
-          rel="noopener noreferrer"
-        >
-          <Image
-            aria-hidden
-            src="/file.svg"
-            alt="File icon"
-            width={16}
-            height={16}
-          />
-          Learn
-        </a>
-        <a
-          className="flex items-center gap-2 hover:underline hover:underline-offset-4"
-          href="https://vercel.com/templates?framework=next.js&utm_source=create-next-app&utm_medium=appdir-template-tw&utm_campaign=create-next-app"
-          target="_blank"
-          rel="noopener noreferrer"
-        >
-          <Image
-            aria-hidden
-            src="/window.svg"
-            alt="Window icon"
-            width={16}
-            height={16}
-          />
-          Examples
-        </a>
-        <a
-          className="flex items-center gap-2 hover:underline hover:underline-offset-4"
-          href="https://nextjs.org?utm_source=create-next-app&utm_medium=appdir-template-tw&utm_campaign=create-next-app"
-          target="_blank"
-          rel="noopener noreferrer"
-        >
-          <Image
-            aria-hidden
-            src="/globe.svg"
-            alt="Globe icon"
-            width={16}
-            height={16}
-          />
-          Go to nextjs.org →
-        </a>
-      </footer>
+
+        <div className={styles.logo}>
+          <span className={styles.accent2}>We</span>Verif<span className={styles.accent}>AI</span>
+        </div>
+
+        <div className={styles.navright}>
+          <div className={styles.toggleWrapper}>
+            <label className={styles.switch}>
+              <input 
+                type="checkbox" 
+                checked={darkMode} 
+                onChange={() => setDarkMode(!darkMode)} 
+              />
+              <span className={styles.slider}></span>
+            </label>
+          </div>
+          <button className={styles.createPostBtn} onClick={handleCreatePostClick}>
+            <FaPlus /> UVerifAI
+          </button>
+          <a href="#" className={styles.navright}>Forum</a>
+          <a href="#" className={styles.navright}>Support</a>
+          <button className={styles.signupBtn} onClick={handleSignUpClick}>Sign Up</button>
+          
+        </div>
+      </header>
+
+      <section className={styles.forum}>
+        <h1 className={styles.forumTitle}>Recent Discussions</h1>
+        <div className={styles.posts}>
+          {posts.map((post) => (
+            <div key={post.id} className={styles.post}>
+              <div className={styles.postHeader}>
+                <FaUserCircle className={styles.userIcon} />
+                <div className={styles.postInfo}>
+                  <h2 className={styles.postTitle}>{post.title}</h2>
+                  <p className={styles.postMeta}>Posted by {post.author} · {post.timePosted}</p>
+                </div>
+              </div>
+              <p className={styles.postContent}>{post.content}</p>
+              <div className={styles.postFooter}>
+                <FaCommentAlt className={styles.commentIcon} />
+                <span className={styles.commentCount}>{post.commentCount} comments</span>
+              </div>
+            </div>
+          ))}
+        </div>
+      </section>
+
+      <footer className={styles.footer}>© 2025 ForumName. All rights reserved.</footer>
     </div>
   );
 }
