+++ src/app/post/login.module.css
@@ -8,24 +8,23 @@
     color: #333;
     font-family: "Inter", sans-serif;
     transition: background 0.3s ease, color 0.3s ease;
-  }
+}
 
-  .accent {
+.accent {
     color: #f36100 !important;
-  }
+}
 
-  .accent2 {
+.accent2 {
     color: #007bff !important;
-  }
-  
-  /* DARK MODE */
-  .dark {
+}
+
+/* DARK MODE */
+.dark {
     background: #181818;
     color: #eee;
-  }
-  
+}
 
-  .header {
+.header {
     width: 100%;
     display: flex;
     justify-content: space-between;
@@ -37,16 +36,14 @@
     top: 0;
     z-index: 100;
     transition: background 0.3s ease;
-  }
+}
 
-  
-  .dark .header {
+.dark .header {
     background: #222;
     box-shadow: 0 4px 10px rgba(255, 255, 255, 0.05);
-  }
-  
-  
-  .logo {
+}
+
+.logo {
     position: absolute;
     left: 50%;
     transform: translateX(-50%);
@@ -55,34 +52,32 @@
     font-size: 1.8rem;
     font-weight: 600;
     color: #000;
-  }
-  
-  .dark .logo {
+}
+
+.dark .logo {
     color: #eee;
-  }
-  
-  
-  .toggleWrapper {
+}
+
+.toggleWrapper {
     display: flex;
     align-items: center;
     gap: 5px;
-  }
-  
-  
-  .switch {
+}
+
+.switch {
     position: relative;
     display: inline-block;
     width: 40px;
     height: 20px;
-  }
-  
-  .switch input {
+}
+
+.switch input {
     opacity: 0;
     width: 0;
     height: 0;
-  }
-  
-  .slider {
+}
+
+.slider {
     position: absolute;
     cursor: pointer;
     top: 0;
@@ -92,9 +87,9 @@
     background-color: #007bff;
     transition: 0.4s;
     border-radius: 20px;
-  }
-  
-  .slider:before {
+}
+
+.slider:before {
     position: absolute;
     content: "";
     height: 16px;
@@ -104,17 +99,17 @@
     background-color: #ffffff;
     transition: 0.4s;
     border-radius: 50%;
-  }
-  
-  input:checked + .slider {
-    background-color: #f36100;
-  }
-  
-  input:checked + .slider:before {
-    transform: translateX(20px);
-  }
+}
 
-  .navright {
+input:checked + .slider {
+    background-color: #f36100;
+}
+
+input:checked + .slider:before {
+    transform: translateX(20px);
+}
+
+.navright {
     display: flex;
     gap: 20px;
     padding: 5px 5px;
@@ -122,37 +117,34 @@
     text-decoration: none;
     font-size: 1rem;
     transition: 0.3s;
-    
-  }
+}
 
-  .container.dark .navright {
-    color: #bbb
-  }
+.container.dark .navright {
+    color: #bbb;
+}
 
-  .navItem {
+.navItem {
     padding: 8px 16px;
     gap: 10px;
     color: #555;
     text-decoration: none;
     font-size: 1rem;
     transition: 0.3s;
-  }
-  
-  .container.dark .navItem {
+}
+
+.container.dark .navItem {
     color: #bbb;
-  }
-  
-  .navItem:hover {
+}
+
+.navItem:hover {
     color: #007bff;
-  }
-  
-  .container.dark .navItem:hover {
+}
+
+.container.dark .navItem:hover {
     color: #3ea6ff;
-  }
-  
-  
-  
-  .sidebar {
+}
+
+.sidebar {
     position: fixed;
     left: -250px;
     top: 0;
@@ -162,18 +154,17 @@
     box-shadow: 2px 0 10px rgba(0, 0, 0, 0.1);
     padding-top: 60px;
     transition: left 0.3s ease-in-out;
-  }
-  
-  .dark .sidebar {
+}
+
+.dark .sidebar {
     background: #222;
-  }
-  
-  .sidebar.open {
+}
+
+.sidebar.open {
     left: 0;
-  }
-  
-  
-  .closeButton {
+}
+
+.closeButton {
     background: none;
     border: none;
     font-size: 1.5rem;
@@ -182,121 +173,118 @@
     position: absolute;
     top: 15px;
     right: 15px;
-  }
-  
-  .dark .closeButton {
+}
+
+.dark .closeButton {
     color: #bbb;
-  }
-  
-  
-  .forum {
+}
+
+.hero {
     margin-top: 100px;
     text-align: center;
     width: 90%;
     max-width: 800px;
-  }
-  
-  .forumTitle {
-    font-size: 2rem;
+}
+
+.heroTitle {
+    font-size: 2.5rem;
     font-weight: 700;
-  }
-  
-  
-  .posts {
+    color: #222;
+}
+
+.dark .heroTitle {
+    color: #eee;
+}
+
+.heroSubtitle {
+    font-size: 1.2rem;
+    color: #666;
+    margin-top: 10px;
+}
+
+.dark .heroSubtitle {
+    color: #bbb;
+}
+
+.createPostBox {
+    background: #ffffff;
+    border-radius: 12px;
+    box-shadow: 0 10px 15px rgba(0, 0, 0, 0.1);
+    width: 90%;
+    max-width: 800px;
+    transition: background 0.3s ease, box-shadow 0.3s ease;
+    margin-top: 20px;
+    padding: 20px;
+}
+
+.container.dark .createPostBox {
+    background: #222;
+    box-shadow: 0 10px 15px rgba(255, 255, 255, 0.1);
+}
+
+.createPostForm {
     display: flex;
     flex-direction: column;
-    gap: 20px;
-    margin-top: 20px;
-  }
-  
-  .post {
-    background: white;
-    padding: 15px;
-    border-radius: 8px;
-    box-shadow: 0 4px 10px rgba(0, 0, 0, 0.281);
-    transition: transform 0.2s ease;
-  }
-  
-  .dark .post {
-    background: #222;
-    box-shadow: 0 4px 10px rgba(255, 255, 255, 0.05);
-  }
-  
-  .post:hover {
-    transform: translateY(-3px);
-  }
-  
-  
-  .postHeader {
-    display: flex;
-    align-items: center;
     gap: 10px;
-  }
-  
-  .userIcon {
-    font-size: 2rem;
-    color: #007bff;
-  }
-  
-  .postInfo {
-    text-align: left;
-  }
-  
-  .postTitle {
-    font-size: 1.2rem;
-    font-weight: 600;
-  }
-  
-  .postMeta {
-    font-size: 0.9rem;
-    color: #666;
-  }
-  
-  .dark .postMeta {
-    color: #bbb;
-  }
-  
-  
-  .postContent {
-    margin-top: 10px;
+}
+
+.inputField {
+    width: 100%;
+    padding: 12px;
+    border: 1px solid #ddd;
+    border-radius: 8px;
     font-size: 1rem;
-    color: #444;
-  }
-  
-  .dark .postContent {
-    color: #ddd;
-  }
-  
-  
-  .postFooter {
-    display: flex;
-    align-items: center;
-    gap: 5px;
-    margin-top: 10px;
-    color: #555;
-  }
-  
-  .dark .postFooter {
-    color: #bbb;
-  }
-  
-  .commentIcon {
+    background: #fff;
+    color: #333;
+    transition: border-color 0.3s ease;
+}
+
+.container.dark .inputField {
+    background: #333;
+    color: #eee;
+    border: 1px solid #555;
+}
+
+.inputField:focus {
+    border-color: #007bff;
+    outline: none;
+}
+
+.container.dark .inputField:focus {
+    border-color: #3ea6ff;
+}
+
+.submitButton {
+    background: #007bff;
+    color: #fff;
+    padding: 12px;
+    border: none;
+    border-radius: 8px;
     font-size: 1rem;
-  }
-  
-  .commentCount {
-    font-size: 0.9rem;
-  }
-  
-  
-  .footer {
+    cursor: pointer;
+    transition: background 0.3s ease;
+}
+
+.container.dark .submitButton {
+    background: #3ea6ff;
+}
+
+.submitButton:hover {
+    background: #0056b3;
+}
+
+.container.dark .submitButton:hover {
+    background: #1e90ff;
+}
+
+.footer {
     padding: 20px;
     background: white;
     color: #666;
     text-align: center;
-  }
-  
-  .dark .footer {
+}
+
+.dark .footer {
     background: #222;
     color: #bbb;
-  }
+}
